+++ .gitignore
@@ -38,3 +38,7 @@ public/img/quarter/
 
 assets/css/fonts/
 .yarn/
+
+###> phpstan/phpstan ###
+phpstan.neon
+###< phpstan/phpstan ###
