+++ composer.json
@@ -51,6 +51,7 @@
         "voku/stringy": "^6.3"
     },
     "require-dev": {
+        "phpstan/phpstan": "^1.11",
         "phpunit/phpunit": "^9.5",
         "symfony/browser-kit": "^5.3",
         "symfony/css-selector": "^5.3",
