+++ src/Entity/Pages/IssueNotFoundException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Entity\Pages;
+
+use RuntimeException;
+
+class IssueNotFoundException extends RuntimeException
+{
+}
